New Development: No             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: No              Module(s): Module Names change impacts.            Description: Cleaned up some warnings.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@35192 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/nrao/FITS/GBTScanLogReader.h
@@ -96,7 +96,7 @@ public:
 
     // construct this using the indicated ScanLog FITS file
     GBTScanLogReader(const File& scanlog, Int minscan=-1,
-		     Int maxscan=LONG_MAX);
+		     Int maxscan=(Int) LONG_MAX);
 
     ~GBTScanLogReader();
